backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt==5.3.1
Pillow>=10.1
python-dotenv>=1.0
beautifulsoup4>=4.12
requests>=2.31
lxml>=4.9
gunicorn>=21.2
whitenoise>=6.6
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0